NewProject/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
portalocker==2.0.0
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
